requirements.txt
fastapi
uvicorn
requests
pymupdf
pdfminer.six
paddleocr
paddlepaddle
setuptools
python-multipart